Memperbaiki bug
