+++ hospital_management_system/lib/main.dart
@@ -2,31 +2,31 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:provider/provider.dart';
 
 import 'models/akun_model.dart';
 
 import '/services/perfs_service.dart';
 import '/utilities/constants/color.dart';
 
-import 'package:provider/provider.dart';
-
-import 'screens/home%20screen/home_screen.dart';
-import 'screens/login%20screen/login_screen.dart';
-import 'screens/tenaga kesehatan screen/dokter dan perawat/dokter_screen.dart';
-import 'screens/tenaga kesehatan screen/dokter dan perawat/perawat_screen.dart';
-import 'screens/tenaga%20kesehatan%20screen/dokterPerawat_screen.dart';
-import 'screens/pasien screen/pasien_screen.dart';
-import 'screens/rawat screen/dokter perawat screen/rawat_screen.dart';
+import 'views/home%20screen/home_screen.dart';
+import 'views/login%20screen/login_screen.dart';
+import 'views/tenaga kesehatan screen/dokter dan perawat/dokter_screen.dart';
+import 'views/tenaga kesehatan screen/dokter dan perawat/perawat_screen.dart';
+import 'views/tenaga%20kesehatan%20screen/dokterPerawat_screen.dart';
+import 'views/pasien screen/pasien_screen.dart';
+import 'views/rawat screen/dokter perawat screen/rawat_screen.dart';
 
 import 'viewModels/dokter perawat viewModel/dokterPerawat_provider.dart';
-import 'viewModels/home provider/home_provider.dart';
+import 'viewModels/home viewModel/home_provider.dart';
 import 'viewModels/login viewModel/login_provider.dart';
-import 'viewModels/main layout provider/mainLayout_provider.dart';
+import 'viewModels/main layout viewModel/mainLayout_provider.dart';
 import 'viewModels/pasien viewModel/pasien_provider.dart';
 import 'viewModels/rawat viewModel/rawat_provider.dart';
 
-void main() {
+void main() async {
   WidgetsFlutterBinding.ensureInitialized();
+
   SystemChrome.setPreferredOrientations([
     DeviceOrientation.landscapeRight,
     DeviceOrientation.landscapeLeft,
