--- hospital_management_system/lib/utilities/constants/jwt_decode.dart
@@ -1,53 +1,0 @@
-import 'dart:convert';
-
-class JwtDecode {
-  Map<String, dynamic> parseJwt(String token) {
-    final parts = token.split('.');
-    if (parts.length != 3) {
-      throw Exception('invalid token');
-    }
-
-    final payload = _decodeBase64(parts[1]);
-    final payloadMap = json.decode(payload);
-    if (payloadMap is! Map<String, dynamic>) {
-      throw Exception('invalid payload');
-    }
-    return payloadMap;
-  }
-
-  String _decodeBase64(String str) {
-    String output = str.replaceAll('-', '+').replaceAll('_', '/');
-
-    switch (output.length % 4) {
-      case 0:
-        break;
-      case 2:
-        output += '==';
-        break;
-      case 3:
-        output += '=';
-        break;
-      default:
-        throw Exception('Illegal base64url string!"');
-    }
-    return utf8.decode(base64Url.decode(output));
-  }
-
-  bool isExpired(String token) {
-    final DateTime? expirationDate = getExpiryDate(token);
-    if (expirationDate != null) {
-      return DateTime.now().isAfter(expirationDate);
-    } else {
-      return false;
-    }
-  }
-
-  DateTime? getExpiryDate(String token) {
-    final Map<String, dynamic> payload = parseJwt(token);
-    if (payload['exp'] != null) {
-      return DateTime.fromMillisecondsSinceEpoch(0, isUtc: true)
-          .add(Duration(seconds: payload["exp"]));
-    }
-    return null;
-  }
-}
