+++ hospital_management_system/lib/viewModels/pasien viewModel/pasien_provider.dart
@@ -21,7 +21,7 @@ class PasienProvider extends ChangeNotifier {
   TextEditingController searchController = TextEditingController();
 
   PasienProvider() {
-    this.getDataApiPasien();
+    getDataApiPasien();
   }
 
   Future getDataApiPasien() async {
