+++ hospital_management_system/lib/views/home screen/components/content.dart
