+++ hospital_management_system/pubspec.yaml
@@ -21,6 +21,7 @@ dependencies:
   dio: ^4.0.6
   date_format: ^2.0.6
   internet_connection_checker: ^0.0.1+4
+  dart_ping: ^6.1.2
